+++ src/@sdk/react/components/CredentialsProvider/types.ts
@@ -1,0 +1,3 @@
+export interface IProps {
+  children?: React.ReactNode;
+}
