+++ src/@sdk/react/components/SaleorProvider/index.ts
@@ -1,0 +1,1 @@
+export * from "./SaleorProvider";
