+++ src/@sdk/react/components/SaleorProvider/types.ts
@@ -1,0 +1,6 @@
+import ApolloClient from "apollo-client";
+
+export interface IProps<TCacheShape> {
+  children?: React.ReactNode;
+  client: ApolloClient<TCacheShape>;
+}
